Organize AccessKeyTable component
