+++ packages/manager/src/features/ObjectStorage/AccessKeyLanding/AccessKeyTable/AccessKeyTable.tsx
@@ -1,0 +1,102 @@
+import {
+  ObjectStorageKey,
+  RegionS3EndpointAndID,
+} from '@linode/api-v4/lib/object-storage';
+import { APIError } from '@linode/api-v4/lib/types';
+import { styled } from '@mui/material/styles';
+import React, { useState } from 'react';
+
+import { Table } from 'src/components/Table';
+
+import { TableCell } from 'src/components/TableCell';
+import { TableHead } from 'src/components/TableHead';
+import { TableRow } from 'src/components/TableRow';
+
+import { useAccountManagement } from 'src/hooks/useAccountManagement';
+import { useFlags } from 'src/hooks/useFlags';
+import { isFeatureEnabled } from 'src/utilities/accountCapabilities';
+
+import { HostNamesDrawer } from '../HostNamesDrawer';
+import { OpenAccessDrawer } from '../types';
+
+import { AccessKeyTableBody } from './AccessKeyTableBody';
+
+export interface AccessKeyTableProps {
+  data: ObjectStorageKey[] | undefined;
+  error: APIError[] | null | undefined;
+  isLoading: boolean;
+  isRestrictedUser: boolean;
+  openDrawer: OpenAccessDrawer;
+  openRevokeDialog: (objectStorageKey: ObjectStorageKey) => void;
+}
+
+export const AccessKeyTable = (props: AccessKeyTableProps) => {
+  const {
+    data,
+    error,
+    isLoading,
+    isRestrictedUser,
+    openDrawer,
+    openRevokeDialog,
+  } = props;
+
+  const [showHostNamesDrawer, setShowHostNamesDrawers] = useState<boolean>(
+    false
+  );
+  const [hostNames, setHostNames] = useState<RegionS3EndpointAndID[]>([]);
+
+  const flags = useFlags();
+  const { account } = useAccountManagement();
+
+  const isObjMultiClusterEnabled = isFeatureEnabled(
+    'Object Storage Access Key Regions',
+    Boolean(flags.objMultiCluster),
+    account?.capabilities ?? []
+  );
+
+  return (
+    <>
+      <Table
+        aria-label="List of Object Storage Access Keys"
+        colCount={2}
+        data-testid="data-qa-access-key-table"
+        rowCount={data?.length}
+      >
+        <TableHead>
+          <TableRow data-qa-table-head>
+            <StyledLabelCell data-qa-header-label>Label</StyledLabelCell>
+            <StyledLabelCell data-qa-header-key>Access Key</StyledLabelCell>
+            {isObjMultiClusterEnabled && (
+              <StyledLabelCell data-qa-header-key>
+                Regions/S3 Hostnames
+              </StyledLabelCell>
+            )}
+            {/* empty cell for kebab menu */}
+            <TableCell />
+          </TableRow>
+        </TableHead>
+        <AccessKeyTableBody
+          data={data}
+          error={error}
+          isLoading={isLoading}
+          isRestrictedUser={isRestrictedUser}
+          openDrawer={openDrawer}
+          openRevokeDialog={openRevokeDialog}
+          setHostNames={setHostNames}
+          setShowHostNamesDrawers={setShowHostNamesDrawers}
+        />
+      </Table>
+      {isObjMultiClusterEnabled && (
+        <HostNamesDrawer
+          onClose={() => setShowHostNamesDrawers(false)}
+          open={showHostNamesDrawer}
+          regions={hostNames}
+        />
+      )}
+    </>
+  );
+};
+
+const StyledLabelCell = styled(TableCell)(() => ({
+  width: '35%',
+}));
